requirements.txt
textual>=0.85.0
anthropic>=0.39.0
google-generativeai>=0.8.3
python-dotenv>=1.0.1
aiosqlite>=0.20.0
pydantic-settings>=2.6.1
pyyaml>=6.0.2
tiktoken>=0.8.0
tenacity>=9.0.0
